+++ FixedPaletteTool/Runtime/Enums/COLOR.cs
@@ -1,0 +1,10 @@
+﻿namespace FixedColorPaletteTool.Enums
+{
+    public enum COLOR
+    {
+        NONE,
+        PRIMARY,
+        SECONDARY,
+        TERTIARY
+    }
+}